+++ src/HackF5.UnitySpy/IManagedObjectInstance.cs
@@ -1,6 +1,7 @@
 ﻿namespace HackF5.UnitySpy
 {
     using JetBrains.Annotations;
+    using System;
 
     /// <summary>
     /// Represents an object instance in managed memory.
@@ -51,6 +52,6 @@
         /// </returns>
         TValue GetValue<TValue>(string fieldName, string typeFullName);
 
-        uint GetAddress();
+        IntPtr GetAddress();
     }
 }
